reglage document exeption Nul Maj

+++ RSE.NET BACKUP 23-10-18 09h49m/DAL/Models/Document.cs
@@ -12,7 +12,7 @@ namespace DAL.Models {
         public string Description { get; set; }
         public DateTime Date { get; set; }
         public string Lien { get; set; }
-        public float Taille { get; set; }
+        public double Taille { get; set; }
         public string Format { get; set; }
         public int Id_Emp_Creee { get; set; }
         public int? Id_Emp_Maj { get; set; }
